refactor: move to internal package
