+++ internal/testutil/cmd_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/require"
 
-	"github.com/cosmos/cosmos-sdk/testutil"
+	"github.com/cosmos/cosmos-sdk/internal/testutil"
 )
 
 // TestSetArgsWithOriginalMethod is used to illustrate cobra.Command.SetArgs won't reset args as expected
